Adds config file for connection strings to Mongo

+++ database_driver/mongodb_driver.py
@@ -2,15 +2,18 @@ from pymongo import MongoClient, errors
 from bson.json_util import dumps
 import json
 
+# Load Config file
+with open('config/mongo_config.json', 'r') as file:
+    config_data = json.load(file)
+
 try:
     # Connect to MongoDB (use the default host and port)
-    client = MongoClient('mongodb://localhost:27017/')
-    db = client['RecipeDatabase']
+    client = MongoClient(config_data["mongo-client-host"])
+    db = client[config_data["database-name"]]
 
     # Set up all collection objects
-    pantry_essentials_collection = db['pantryEssentials']
-    recipe_link_collection = db['recipeLinks']
-    saved_recipe_collection = db['savedRecipes']
+    pantry_essentials_collection = db[config_data["pantry-essentials-collection-name"]]
+    saved_recipe_collection = db[config_data["saved-recipes-collection-name"]]
     print("Successfully connected to MongoDB database")
 except Exception as e:
     print(f"MongoDB Error: {str(e)}")
@@ -21,6 +24,9 @@ def insert_pantry_essentials(ingredient_list: dict):
     Inserts an ingredient list to be stored as pantry essentials
     :param ingredient_list: dict containing list of ingredients
     """
+    if type(ingredient_list) is not dict:
+        raise TypeError(f"Expected parameter to be of type 'dict' but got {type(ingredient_list).__name__}")
+
     try:
         # Delete existing item and add new item
         pantry_essentials_collection.delete_many({})
@@ -51,6 +57,4 @@ if __name__ == "__main__":
     }
 
     insert_pantry_essentials(ingredients_list)
-
     pantry_essentials = get_pantry_essentials()
-    print(type(pantry_essentials))
